Meta progression

+++ Assets/SoaringHorse/Scripts/Data/Config/ProgressionConfig.cs
@@ -3,9 +3,9 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Soaring Horse/Balance/Progression Config", fileName = "ProgressionConfig")]
 public class ProgressionConfig : ScriptableObject
 {
-    [SerializeField] private float _scoreIncreasePerTick = 1;
+    [SerializeField] private int _scoreIncreasePerTick = 1;
     [SerializeField] private float _scoreTickTime = 3;
 
-    public float ScoreIncreasePerTick { get=> _scoreIncreasePerTick; }
+    public int ScoreIncreasePerTick { get=> _scoreIncreasePerTick; }
     public float ScoreTickTime { get=> _scoreTickTime; }
 }
